# swaggerMoi.json
{
  "swagger": "2.0",
  "info": {
    "description": "save the prefer books of users",
    "version": "1.0.0",
    "title": "Users and Prefer Books API",
    "contact": {
      "email": "moisesparra82@hotmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:8080",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
    "/":{
      "get": {
        "tags": [
          "id"
        ],
        "summary": "Get All the books in database",
        "operationId": "getBookById",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "invalid title"
          }
        }
      }
    },
...
  },
...
}